refactor: change name gateway to client
