+++ Sources/Interfaces/Client/IBasaltConnectionManager.ts
