+++ Sources/Interfaces/Client/IBasaltSocketClient.ts
@@ -1,0 +1,3 @@
+import { IBasaltConnectionManager, IBasaltEventListenerManager } from '@/Interfaces';
+
+export interface IBasaltSocketClient extends IBasaltConnectionManager, IBasaltEventListenerManager {}
