--- Sources/Interfaces/Gateway/IBasaltSocketGateway.ts
@@ -1,3 +1,0 @@
-import { IBasaltConnectionManager, IBasaltEventListenerManager } from '@/Interfaces';
-
-export interface IBasaltSocketGateway extends IBasaltConnectionManager, IBasaltEventListenerManager {}
