+++ Sources/Interfaces/index.ts
@@ -5,7 +5,7 @@ export * from './Server/IBasaltWebSocketBehavior';
 export * from './Server/IBasaltSocketEvents';
 export * from './Server/IBasaltHttpResponse';
 export * from './Server/IBasaltHttpRequest';
-export * from './Gateway/IBasaltSocketGatewayOptions';
-export * from './Gateway/IBasaltConnectionManager';
-export * from './Gateway/IBasaltSocketGateway';
-export * from './Gateway/IBasaltEventListenerManager';
+export * from './Client/IBasaltSocketClientOptions';
+export * from './Client/IBasaltConnectionManager';
+export * from './Client/IBasaltSocketClient';
+export * from './Client/IBasaltEventListenerManager';
